+++ pkg/webhooks/proxypod/proxypod.go
@@ -18,99 +18,41 @@ package proxypod // import "admiralty.io/multicluster-scheduler/pkg/webhooks/pro
 
 import (
 	"context"
-	"fmt"
 	"net/http"
-	"os"
 
 	"admiralty.io/multicluster-scheduler/pkg/common"
-	"github.com/ghodss/yaml"
-	admissionregistrationv1beta1 "k8s.io/api/admissionregistration/v1beta1"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/json"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/runtime/inject"
-	"sigs.k8s.io/controller-runtime/pkg/webhook"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/admission/builder"
-	atypes "sigs.k8s.io/controller-runtime/pkg/webhook/admission/types"
+	"sigs.k8s.io/yaml"
 )
 
-var webhookName = "proxypod.multicluster.admiralty.io"
-
-func NewServer(mgr manager.Manager, namespace string) (*webhook.Server, error) {
-	w, err := NewWebhook(mgr)
-	if err != nil {
-		return nil, fmt.Errorf("cannot build webhook: %v", err)
-	}
-
-	deployName := os.Getenv("DEPLOYMENT_NAME")
-
-	s, err := webhook.NewServer(deployName, mgr, webhook.ServerOptions{
-		Port:    9876, // TODO debug why cannot default to 443
-		CertDir: "/tmp/cert",
-		BootstrapOptions: &webhook.BootstrapOptions{
-			Secret: &types.NamespacedName{
-				Namespace: namespace,
-				Name:      deployName + "-cert",
-			},
-			Service: &webhook.Service{
-				Namespace: namespace,
-				Name:      deployName,
-				// Selectors should select the pods that runs this webhook server.
-				Selectors: map[string]string{
-					"app.kubernetes.io/name": deployName,
-					// HACK: there should be more here (cf. selectorLabels in helm chart)
-					// we could get the labels using the downward API, just like we get DEPLOYMENT_NAME
-					// but this won't be necessary once we stop using BootstrapOptions
-				},
-			},
-		},
-	})
-	if err != nil {
-		return nil, fmt.Errorf("cannot create server: %v", err)
-	}
-
-	if err := s.Register(w); err != nil {
-		return nil, fmt.Errorf("cannot register webhook with server: %v", err)
-	}
-
-	return s, nil
-}
-
-// https://kubernetes.slack.com/archives/CAR30FCJZ/p1547254570666900
-func NewWebhook(mgr manager.Manager) (*admission.Webhook, error) {
-	return builder.NewWebhookBuilder().
-		Name(webhookName).
-		Mutating().
-		Operations(admissionregistrationv1beta1.Create). // TODO: update (but careful not to proxy the proxy)
-		WithManager(mgr).
-		ForType(&corev1.Pod{}).
-		Handlers(&Handler{mutator: mutator{}}).
-		FailurePolicy(admissionregistrationv1beta1.Fail).
-		NamespaceSelector(&metav1.LabelSelector{MatchLabels: map[string]string{"multicluster-scheduler": "enabled"}}).
-		Build()
-}
-
 type Handler struct {
-	decoder atypes.Decoder
+	decoder *admission.Decoder
 	client  client.Client
 	mutator mutator
 }
 
-func (h *Handler) Handle(ctx context.Context, req atypes.Request) atypes.Response {
+func (h *Handler) Handle(_ context.Context, req admission.Request) admission.Response {
 	srcPod := &corev1.Pod{}
 	err := h.decoder.Decode(req, srcPod)
 	if err != nil {
-		return admission.ErrorResponse(http.StatusBadRequest, err)
+		return admission.Errored(http.StatusBadRequest, err)
 	}
 
 	proxyPod := srcPod.DeepCopy()
 	if err := h.mutator.mutate(proxyPod); err != nil {
-		return admission.ErrorResponse(http.StatusInternalServerError, err)
+		return admission.Errored(http.StatusInternalServerError, err)
 	}
-	return admission.PatchResponse(srcPod, proxyPod)
+
+	proxyPodRaw, err := json.Marshal(proxyPod)
+	if err != nil {
+		return admission.Errored(http.StatusInternalServerError, err)
+	}
+
+	return admission.PatchResponseFromRaw(req.Object.Raw, proxyPodRaw)
 }
 
 type mutator struct {
@@ -122,29 +64,35 @@ func (m mutator) mutate(pod *corev1.Pod) error {
 		return nil
 	}
 
-	srcPodManifest, err := yaml.Marshal(pod)
-	if err != nil {
-		return err
+	// only save the source manifest if it's not set already
+	// webhooks may be run multiple times on the same object
+	// and have to be idempotent
+	// if we didn't check, we could lose the source scheduling constraints that we remove below
+	if _, ok := pod.Annotations[common.AnnotationKeySourcePodManifest]; !ok {
+		srcPodManifest, err := yaml.Marshal(pod)
+		if err != nil {
+			return err
+		}
+
+		// pod.Annotations is not nil because we checked it contains AnnotationKeyElect
+		pod.Annotations[common.AnnotationKeySourcePodManifest] = string(srcPodManifest)
 	}
 
-	// pod.Annotations is not nil because we checked it contains AnnotationKeyElect
-	pod.Annotations[common.AnnotationKeySourcePodManifest] = string(srcPodManifest)
+	pod.Spec.Tolerations = []corev1.Toleration{{
+		Key:   "virtual-kubelet.io/provider",
+		Value: "admiralty",
+	}}
 
-	for i, c := range pod.Spec.Containers {
-		pod.Spec.Containers[i] = corev1.Container{
-			Name:    c.Name,
-			Image:   image,
-			Command: command}
-		// the feedback controller will send SIGUSR1 or SIGUSR2 when the delegate pod succeeds or fails, resp.
-	}
-	// TODO: add resource reqs/lims + other best practices
-	// TODO!!! remove scheduling prefs
+	pod.Spec.NodeName = "admiralty"
+
+	// remove other scheduling constraints (will be respected in target cluster, from source pod manifest)
+	pod.Spec.SchedulerName = ""
+	pod.Spec.NodeSelector = nil
+	pod.Spec.Affinity = nil
+	pod.Spec.TopologySpreadConstraints = nil
 
 	return nil
 }
-
-var image = "busybox"
-var command = []string{"sh", "-c", "trap 'exit 0' SIGUSR1; trap 'exit 1' SIGUSR2; (while sleep 3600; do :; done) & wait"}
 
 // Handler implements inject.Client.
 // A client will be automatically injected.
@@ -158,10 +106,10 @@ func (h *Handler) InjectClient(c client.Client) error {
 
 // Handler implements inject.Decoder.
 // A decoder will be automatically injected.
-var _ inject.Decoder = &Handler{}
+var _ admission.DecoderInjector = &Handler{}
 
 // InjectDecoder injects the decoder.
-func (h *Handler) InjectDecoder(d atypes.Decoder) error {
+func (h *Handler) InjectDecoder(d *admission.Decoder) error {
 	h.decoder = d
 	return nil
 }