+++ release/release.sh
@@ -5,7 +5,6 @@ VERSION="$1"
 
 IMAGES=(
   "multicluster-scheduler-agent"
-  "multicluster-scheduler-pod-admission-controller"
   "multicluster-scheduler-basic"
   "multicluster-scheduler-remove-finalizers"
 )
@@ -14,5 +13,9 @@ for IMAGE in "${IMAGES[@]}"; do
   docker push "quay.io/admiralty/$IMAGE:$VERSION"
 done
 
-# TODO: upload Helm chart
+helm package charts/multicluster-scheduler -d _out
+cp charts/index.yaml _out/
+helm repo index _out --merge _out/index.yaml --url https://charts.admiralty.io
+
+# TODO: upload Helm chart and new index
 # TODO: also tag images with latest