+++ test/e2e/single-namespace/test.sh
@@ -2,6 +2,7 @@
 set -euo pipefail
 
 source test/e2e/aliases.sh
+source test/e2e/webhook_ready.sh
 
 setup() {
   helm1 upgrade --install multicluster-scheduler charts/multicluster-scheduler -f test/e2e/single-namespace/values-cluster1.yaml
@@ -9,6 +10,9 @@ setup() {
 
   ./kubemcsa export --kubeconfig kubeconfig-cluster1 c1 --as remote | k1 apply -f -
   ./kubemcsa export --kubeconfig kubeconfig-cluster1 c2 --as remote | k2 apply -f -
+
+  webhook_ready 1 default multicluster-scheduler-agent multicluster-scheduler-agent multicluster-scheduler-agent-cert
+  webhook_ready 2 default multicluster-scheduler-agent multicluster-scheduler-agent multicluster-scheduler-agent-cert
 }
 
 tear_down() {